split token from class
