+++ src/packages/core/tree/default/index.ts
@@ -1,2 +1,3 @@
 export { UmbDefaultTreeElement } from './default-tree.element.js';
 export { UmbDefaultTreeContext } from './default-tree.context.js';
+export { UMB_TREE_CONTEXT } from './default-tree.context-token.js';
